# docs/openapi.json
{"openapi": "3.1.0", "info": {"title": "Car Rent", "description": "A simple api to rent a car", "version": "0.1"}, "paths": {"/cars": {"get": {"summary": "List Cars", "description": "Endpoint that returns information about all the cars.\nArgs:\n    Not availible\nReturns:\n    The list of all cars", "operationId": "list_cars_cars_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/Car_Out"}, "type": "array", "title": "Response List Cars Cars Get"}}}}}}, "post": {"summary": "Add Car", "description": "Endpoint to add cars in database\nArgs:\n    Name,Model,Category,Year,Price,Rate\nReturns:\n    Return the result of add!", "operationId": "add_car_cars_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Car_In"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Car_Out"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/remove_cars": {"post": {"summary": "Remove Car", "description": "Endpoint to remove car\nArgs:\n    Name\nReturns:\n    Return the result of remove!", "operationId": "remove_car_remove_cars_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Car_In"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Car_Out"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"Car_In": {"properties": {"name": {"type": "string", "title": "Name"}, "model": {"type": "string", "title": "Model"}, "category": {"type": "string", "title": "Category"}, "year": {"type": "integer", "title": "Year"}, "price": {"type": "integer", "title": "Price"}, "rate": {"type": "integer", "title": "Rate"}, "date": {"type": "string", "format": "date-time", "title": "Date"}}, "type": "object", "required": ["name", "model", "category", "year", "price", "rate", "date"], "title": "Car_In"}, "Car_Out": {"properties": {"name": {"type": "string", "title": "Name"}, "model": {"type": "string", "title": "Model"}, "category": {"type": "string", "title": "Category"}, "year": {"type": "integer", "title": "Year"}, "price": {"type": "integer", "title": "Price"}, "rate": {"type": "integer", "title": "Rate"}, "date": {"type": "string", "format": "date-time", "title": "Date"}}, "type": "object", "required": ["name", "model", "category", "year", "price", "rate", "date"], "title": "Car_Out"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}